chore: #94 move block validation to separate file
